pyproject.toml
[build-system]
build-backend = "pdm.backend"
requires = ["pdm-backend"]

[project]
authors = [
  {name = "Tohrusky", email = "65994850+Tohrusky@users.noreply.github.com"}
]
classifiers = [
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12"
]
dependencies = [
  "realcugan-ncnn-py>=2.0.0",
  "realesrgan-ncnn-py>=2.0.0",
  "waifu2x-ncnn-py>=2.0.0",
  "PyYAML~=6.0",
  "loguru~=0.7.0"
]
license = {text = "BSD-3-Clause"}
name = "Final2x-core"
readme = "README.md"
requires-python = ">=3.8"
version = "2.0.0"

[project.urls]
Homepage = "https://github.com/Final2x/Final2x-core"
Repository = "https://github.com/Final2x/Final2x-core"

[tool.coverage.run]
omit = ["src"]

[tool.mypy]
disable_error_code = "attr-defined"
disallow_any_generics = false
ignore_missing_imports = true
strict = true
warn_return_any = false

[tool.pdm.dev-dependencies]
lint = [
  "pre-commit",
  "ruff",
  "mypy"
]
test = [
  "pytest",
  "pytest-cov",
  "scikit-image"
]

[tool.pdm.scripts.lint]
composite = [
  "mypy src tests",
  "ruff check . --fix --exit-non-zero-on-fix",
  "ruff format ."
]
help = "Check code style against linters"

[tool.pdm.scripts.post_install]
composite = [
  "pre-commit install"
]
help = "Install the pre-commit hook"

[tool.pdm.scripts.test]
cmd = "pytest tests --cov=src --cov-report=xml --cov-report=html"
help = "Run tests with coverage"

[tool.ruff]
extend-ignore = ["B018", "B019"]
extend-select = [
  "I",  # isort
  "B",  # flake8-bugbear
  "C4",  # flake8-comprehensions
  "PGH",  # pygrep-hooks
  "RUF",  # ruff
  "W",  # pycodestyle
  "YTT"  # flake8-2020
]
fixable = ["ALL"]
line-length = 120

[tool.ruff.format]
indent-style = "space"
line-ending = "auto"
quote-style = "double"
skip-magic-trailing-comma = false

[tool.ruff.isort]
combine-as-imports = true
known-first-party = ["pdm"]

[tool.ruff.mccabe]
max-complexity = 10
